Add a small example that moves functions.
It moves functions from one bitcode file to another. It works only on a single file at the moment
and will require more work but this is the first promising result. The function `llvm::CloneModule`
was used for a reference as it provides a good reference for moving everything. It also shows that
the code can be used to split the modules into multiple different modules, and then into different
shared objects respectively. The GlobalValue moving code works fine it turns out but it doesn't
clone them 100% identically as there are still some differences.