Functions are filtered based on moveability.
All of the functions that cannot, be moved are not printed. The
functions that can be moved are the internal functions that are the only
internal functions referenced by their called and don't reference an
internal function and are not referenced by store instructions (no
function pointers). The program also correctly identifies all of the
things that need to be moved such as global variable references.