Support for splitting with global variables.
This work was needed as global variables are more susceptible to
changing their linkage and visibility as they also need ExternalLinkage
and DefaultVisibility to be movable outside of their module. That
however introduced a linker error in some cases: `relocation R_X86_64_PC32
against symbol`. This relocation should not have been used with static
global variables. This is why the globals first need to be modified.
So that they are truly global, and then modified again to be externally
initialized as to when they are moved with their respective function
users the initialization code will be in a different module.

The code only works when the global variable's initializer doesn't
depend on other GlobalVariables or Constants. To make this work the
semantics of the globals needs to be changed. Splitting CPython using
this method still doesn't work.