Move globals and their dependencies properly.
This commit finally allows real world programs to work. Whilst
still not finished the current state of the program is on a
competent enough level that, most programs should work sufficiently
well. While still not working with CPython, it works with Lua
and no tests are failing providing the original Lua is
build properly.

The main reason behind why these things were unable to be achieved
before is the fact that by moving a global variable the things used
when initializing that global variable were not moved with it. To
solve that problem the entire `llvm::User` graph is walked and thus
all the dependencies resolved. Then the entire dependency graph is
moved.