Fix visibility issues with global variables
Since the splitter operates by creating a shared object for each fun-
ction and global variable/constant, it is crucial to make them visible
outside each shared object. For example, if a shared object is using
a variable defined in some other shared object, that variable should
be visible. In fact, when all the shared objects are linked dynamically
to create the final executable program, if some of the symbols are not
visible, the linking process will fail.
Without the changes provided in this commit constants would be exported
with hidden visibility, making them not visible outside their own shared
object.
This commit also removes some redundant code, discovered while making
the aforemetioned changes, and so improves general readibility.