Update README with more detailed build instructions.
This adds the LLVM minimum version requirement and a few sentences
about how to build the splitter.

This also updates the `llvm-config` invocation to not use
`--link-shared` (using `--link-shared` causes the build to fail as
`libLLVM.so` is not built by default). Alternatively, we could have
changed the `cheribuild` invocation to build LLVM with
`-DLLVM_BUILD_LLVM_DYLIB=1` and `-DLLVM_LINK_LLVM_DYLIB=1`.